+++ test/Spec.hs
@@ -119,7 +119,7 @@ spec = do
             (_, Just hosthome) -> do
               assertUserSwitch out user
               out `shouldContain` ("-e=HOME=" ++ hosthome)
-              out `shouldContain` ("--workdir " ++ hosthome)
+              out `shouldNotContain` ("--workdir " ++ hosthome)
               out `shouldContain` "--passwd-entry"
               out `shouldContain` (":" ++ hosthome ++ ":/bin/sh")
             (_, Nothing) ->